requirements.txt
huggingface_hub>=0.21.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
stability-sdk>=0.8.4
Pillow>=9.5.0 